Impl database for storage user info

+++ app/src/main/java/com/example/shedulebirthday/view/ItemClickListener.kt
@@ -1,6 +1,6 @@
 package com.example.shedulebirthday.view
 
 interface ItemClickListener {
-    fun onCellClickListener(idItem: Int)
-    fun onCellLongClickListener(idItem: Int)
+    fun onCellClickListener(idItem: Long)
+    fun onCellLongClickListener(idItem: Long)
 }
